WebSocket Split and refactor
